requirements.txt
scrapy
camelot-py[base]
